+++ FactoryMethodDemo/FactoryMethodDemo/ConcreteImpl/MoneyBack.cs
@@ -1,0 +1,25 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace FactoryMethodDemo.ConcreteImpl
+{
+    //
+    public class MoneyBack : ICreditCard
+    {
+        public int GetAnnualCharge()
+        {
+            return 1000;
+        }
+
+        public string GetCardType()
+        {
+            return "MoneyBack Credit Card";
+        }
+
+        public int GetCreditLimit()
+        {
+            return 100000;
+        }
+    }
+}
