+++ FactoryMethodDemo/FactoryMethodDemo/Enums/CreditCardTypes.cs
@@ -1,0 +1,13 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace FactoryMethodDemo.Enums
+{
+    public enum CreditCardTypes
+    {
+        MoneyBack,
+        Platinum,
+        Titanium
+    }
+}
