+++ FactoryMethodDemo/FactoryMethodDemo/Interface/ICreditCard.cs
@@ -1,0 +1,15 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace FactoryMethodDemo
+{
+    //This Product class
+    public interface ICreditCard
+    {
+        string GetCardType();
+        int GetCreditLimit();
+        int GetAnnualCharge();
+
+    }
+}
