--- FactoryMethodDemo/FactoryMethodDemo/Product/ICreditCard.cs
@@ -1,15 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace FactoryMethodDemo
-{
-    //This Product class
-    public interface ICreditCard
-    {
-        string GetCardType();
-        int GetCreditLimit();
-        int GetAnnualCharge();
-
-    }
-}
